+++ docs/src/py_spec_subscriber.sh
@@ -1,0 +1,1 @@
+PYTHONPATH=. asyncapi-subscriber --api-module py_spec_user_events